Vision1/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
distlib==0.3.6
Django==4.0.6
django-crispy-forms==1.14.0
django-formtools==2.3
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
filelock==3.8.0
fonttools==4.37.4
kiwisolver==1.4.4
kombu==5.2.4
matplotlib==3.6.1
mysqlclient==2.1.1
numpy==1.23.2
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
prompt-toolkit==3.0.30
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tenacity==8.1.0
tzdata==2022.1
vine==5.0.0
virtualenv==20.16.4
wcwidth==0.2.5
